# Retro-Forum-Hub/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HiddenFreeways API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /threads:
    get:
      operationId: listThreads
      tags: [threads]
      summary: List threads with optional filters
      parameters:
        - name: categorySlug
          in: query
          required: false
          schema:
            type: string
        - name: locationId
          in: query
          required: false
          schema:
            type: integer
        - name: stateSlug
          in: query
          required: false
          schema:
            type: string
        - name: pinnedOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of threads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreadSummary"
    post:
      operationId: createThread
      tags: [threads]
      summary: Start a new thread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThreadBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thread"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /threads/{id}:
    get:
      operationId: getThread
      tags: [threads]
      summary: Get a thread with its posts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Thread detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreadDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /threads/{id}/posts:
    post:
      operationId: createPost
      tags: [posts]
      summary: Reply to a thread
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ------------------- FEED -------------------
...
  /threads/{id}/vote:
    post:
      operationId: voteOnThread
      tags: [votes]
      summary: Up/downvote (or clear vote on) a thread
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteBody"
      responses:
        "200":
          description: Updated vote totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /posts/{id}/vote:
    post:
      operationId: voteOnPost
      tags: [votes]
      summary: Up/downvote (or clear vote on) a post
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoteBody"
      responses:
        "200":
          description: Updated vote totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ------------------- CHAT ROOMS -------------------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ---------- AUTH ----------
...
    Thread:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        categoryId:
          type: integer
        categorySlug:
          type: string
        categoryName:
          type: string
        locationId:
          type: ["integer", "null"]
        locationName:
          type: ["string", "null"]
        stateId:
          type: ["integer", "null"]
        stateSlug:
          type: ["string", "null"]
        authorId:
          type: integer
        authorUsername:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        isPinned:
          type: boolean
        isLocked:
          type: boolean
        viewCount:
          type: integer
        replyCount:
          type: integer
        lastActivityAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - body
        - categoryId
        - categorySlug
        - categoryName
        - locationId
        - locationName
        - stateId
        - stateSlug
        - authorId
        - authorUsername
        - authorAvatarUrl
        - isPinned
        - isLocked
        - viewCount
        - replyCount
        - lastActivityAt
        - createdAt
    ThreadSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        excerpt:
          type: string
        categoryId:
          type: integer
        categorySlug:
          type: string
        categoryName:
          type: string
        locationId:
          type: ["integer", "null"]
        locationName:
          type: ["string", "null"]
        stateSlug:
          type: ["string", "null"]
        authorId:
          type: integer
        authorUsername:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        isPinned:
          type: boolean
        isLocked:
          type: boolean
        replyCount:
          type: integer
        lastActivityAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - excerpt
        - categoryId
        - categorySlug
        - categoryName
        - locationId
        - locationName
        - stateSlug
        - authorId
        - authorUsername
        - authorAvatarUrl
        - isPinned
        - isLocked
        - replyCount
        - lastActivityAt
        - createdAt
    ThreadDetail:
      type: object
      properties:
        thread:
          $ref: "#/components/schemas/Thread"
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
      required: [thread, posts]
    CreateThreadBody:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        categoryId:
          type: integer
        locationId:
          type: ["integer", "null"]
      required: [title, body, categoryId]

    # ---------- POSTS ----------
    Post:
      type: object
      properties:
        id:
          type: integer
        threadId:
          type: integer
        body:
          type: string
        authorId:
          type: integer
        authorUsername:
          type: string
        authorAvatarUrl:
          type: ["string", "null"]
        authorRole:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - threadId
        - body
        - authorId
        - authorUsername
        - authorAvatarUrl
        - authorRole
        - createdAt
    CreatePostBody:
      type: object
      properties:
        body:
          type: string
      required: [body]

    # ---------- FEED ----------
...
    VoteBody:
      type: object
      properties:
        value:
          type: string
          enum: [up, down, none]
      required: [value]
    VoteResult:
      type: object
      properties:
        upvotes:
          type: integer
        downvotes:
          type: integer
        score:
          type: integer
        myVote:
          type: string
          enum: [up, down, none]
      required: [upvotes, downvotes, score, myVote]
...